requirements.txt
Django
django-rest-framework
django-widget-tweaks
faker
pytest
pytest-django
